pyproject.toml
[tool.poetry]
name = "dataprocer"
version = "0.1.0"
description = ""
authors = ["Agung Setiaji <mragungsetiaji@gmail.com>"]

[tool.poetry.dependencies]
python = "3.9"
pykg = "^0.1.1"
python-dotenv = "^0.15.0"

[tool.poetry.dev-dependencies]
pytest = "^5.2"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
